Updated the webpack build for prod distribution. Added new package.json dependency to support object rest spread syntax. Fixed imports.
