+++ template/.babelrc
@@ -5,6 +5,7 @@
   ],
   "plugins": [
   	"react-hot-loader/babel",
+  	"syntax-object-rest-spread",
   	"transform-class-properties"
   ]
 }
