--- template/electron/renderer/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>Component</title>
-    <link rel="stylesheet" type="text/css" href="../../lib/styles/compass/index.css">
-    <link rel="stylesheet" type="text/css" href="../../lib/styles/index.css">
-  </head>
-  <body>
-    <div id="container">
-    </div>
-    <script src="index.js" charset="UTF-8" async></script>
-  </body>
-</html>
