+++ template/package.json
@@ -3,28 +3,24 @@
   "productName": "{{capitalcase name}} plugin",
   "version": "0.0.1",
   "description": "{{description}}",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
+    "build": "webpack --config ./config/webpack.prod.config.js",
+    "start": "webpack-dev-server --config ./config/webpack.dev.config.js",
     "ci": "npm run check && npm test",
     "fmt": "mongodb-js-fmt ./*.js ./test/*.js",
     "check": "mongodb-js-precommit './src/**/*{.js,.jsx}' './test/**/*.js' index.js",
-    "compile": "npm run build",
-    "build": "webpack --config ./config/webpack.prod.config.js",
     "link-plugin": "./scripts/link.sh",
     "unlink-plugin": "./scripts/unlink.sh",
-    "prepublish": "npm run compile",
-    "pretest": "npm run compile",
-    "prestart": "npm run compile",
-    "start": "electron ./electron",
+    "prepublish": "npm run build",
+    "pretest": "npm run build",
     "storybook": "start-storybook -p 9001",
     "test": "hadron-build test",
     "test-enzyme": "npm test -- --enzyme",
     "test-main": "npm test -- --main",
     "test-renderer": "npm test -- --renderer",
     "test-unit": "npm test -- --unit",
-    "test-functional": "npm test -- --functional",
-    "start:dev": "webpack-dev-server --config ./config/webpack.dev.config.js",
-    "start:prod": "webpack --config ./config/webpack.prod.config.js && electron ./electron --noDevServer ."
+    "test-functional": "npm test -- --functional"
   },
   "precommit": [
     "check"
@@ -38,6 +34,7 @@
     "babel-core": "^6.24.1",
     "babel-loader": "^6.4.0",
     "babel-preset-react": "^6.24.1",
+    "babel-plugin-syntax-object-rest-spread": "^6.13.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-register": "^6.9.0",
     "babili-webpack-plugin": "^0.0.11",
